% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_df_predictors}
\alias{create_df_predictors}
\title{Expand covariates in multiple permuations. Like expand.grid, but with 1, 2, ..., n variables chosen}
\usage{
create_df_predictors(names, n_vars = 3)
}
\arguments{
\item{names}{The names of the dataframe to be expanded}

\item{n_vars}{Number of variables to be selected, e.g. expand.grid is called with
1:n_vars variables, and then results are combined and sorted to remove duplicates}
}
\value{
an expanded data frame
}
\description{
\code{create_df_predictors} takes a vector of variable names and creates a dataframe of
}
