% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_ml.R
\name{univariate_forecast_ml}
\alias{univariate_forecast_ml}
\title{Forecast recruitment deviations as a function of covariates using random forest and lasso models}
\usage{
univariate_forecast_ml(
  response,
  predictors,
  model_type,
  n_forecast = 10,
  n_years_ahead = 1,
  control = list(alpha = seq(0.1, 0.9, by = 0.05), lambda = seq(0, 2, by = 0.05), ntree
    = seq(300, 2000, by = 100), mtry = seq(2, 10))
)
}
\arguments{
\item{response}{A data frame of responses for the modeling (values to be forecast) containing
a "time" column and "dev" column}

\item{predictors}{A data frame of predictors used for forecasting recruitment}

\item{model_type}{The type of model used to link predictors to forecasted recruitment. Can
be "randomForest", "glmnet"}

\item{n_forecast}{How many years to use as a holdout / test set}

\item{n_years_ahead}{How many years ahead to forecast (defaults to 1)
1:n_vars variables, and then results are combined and sorted to remove duplicates}

\item{control}{a list of the range and steps of tuning parameters to perform a grid
search over. This includes alpha (defaults from 0.1 to 0.9, by steps of 0.05) and
lambda (defaults from 0 to 2, by steps of 0.05) for glmnet; for randomForest this
includes ntree (defaults from 300 to 2000, by steps of 100) and
mtry (defaults from 2 to 10, by steps of 1)}
}
\value{
a list containing predictions, with elements
\itemize{
\item \code{pred}: the predictions
\item \code{vars}: the variable values used to fit the models
\item \code{coefs}: tidy summaries from each year:iteration. Zeros not included for glmnet
}

#' @examples
response <- data.frame(time = 1:40, dev = rnorm(40))

predictors <- matrix(rnorm(400), ncol = 10)
#colnames(predictors) = paste0("X",1:ncol(predictors))
predictors <- as.data.frame(predictors)
predictors$time <- 1:40
lm_example <- univariate_forecast_ml(response,
predictors,
model_type = "glmnet",
n_forecast = 10,
n_years_ahead = 1)
}
\description{
\code{univariate_forecast_ml} vector of responses and dataframe of predictors and automates
}
