% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate.R
\name{univariate_forecast}
\alias{univariate_forecast}
\title{Forecast recruitment deviations as a function of covariates using GLMs and GAMs}
\usage{
univariate_forecast(
  response,
  predictors,
  model_type,
  n_forecast = 10,
  n_years_ahead = 1,
  max_vars = 3
)
}
\arguments{
\item{response}{A data frame of responses for the modeling (values to be forecast) containing
a "time" column and "dev" column}

\item{predictors}{A data frame of predictors used for forecasting recruitment}

\item{model_type}{The type of model used to link predictors to forecasted recruitment. Can
be "lm", "gam",}

\item{n_forecast}{How many years to use as a holdout / test set}

\item{n_years_ahead}{How many years ahead to forecast (defaults to 1)
1:n_vars variables, and then results are combined and sorted to remove duplicates}

\item{max_vars}{The maximum number of variables to include as predictors; defaults to 3}
}
\value{
a list containing predictions, with elements
\itemize{
\item \code{pred}: the predictions
\item \code{vars}: the variable values used to fit the models
\item \code{coefs}: coefficient estimates from each year:iteration
}
}
\description{
\code{univariate_forecast} vector of responses and dataframe of predictors and automates
}
\examples{
response <- data.frame(time = 1:40, dev = rnorm(40))

predictors <- matrix(rnorm(400), ncol = 10)
#colnames(predictors) = paste0("X",1:ncol(predictors))
predictors <- as.data.frame(predictors)
predictors$time <- 1:40
lm_example <- univariate_forecast(response,
     predictors,
     model_type = "lm",
     n_forecast = 10,
     n_years_ahead = 1,
     max_vars = 3)

}
